+++ experimental/Experimental.jl
@@ -6,8 +6,6 @@ include("InvariantTheory.jl")
 include("GITFans.jl")
 include("GModule.jl")
 
-include("Permutations/Permutations.jl")
-
 include("Schemes/AffineSchemes.jl")
 include("Schemes/SpecOpen.jl")
 include("Schemes/Glueing.jl")
